+++ cloud/acp-spring-cloud-starter/src/main/kotlin/pers/acp/spring/cloud/AcpCloudToolsAutoConfiguration.kt
@@ -1,0 +1,16 @@
+package pers.acp.spring.cloud
+
+import org.springframework.cloud.commons.util.InetUtils
+import org.springframework.context.annotation.*
+import pers.acp.spring.boot.component.ServerTools
+import pers.acp.spring.cloud.component.CloudTools
+
+/**
+ * @author zhangbin by 2018-3-14 15:13
+ * @since JDK 11
+ */
+@Configuration(proxyBeanMethods = false)
+class AcpCloudToolsAutoConfiguration {
+    @Bean
+    fun cloudIpTools(serverTools: ServerTools, inetUtils: InetUtils) = CloudTools(serverTools, inetUtils)
+}